+++ compiler-core/src/type_/printer.rs
@@ -421,6 +421,7 @@ fn test_local_type() {
         module: "mod".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "Cat");
@@ -438,6 +439,7 @@ fn test_prelude_type() {
         module: "gleam".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "Int");
@@ -458,6 +460,7 @@ fn test_shadowed_prelude_type() {
         module: "gleam".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "gleam.Int");
@@ -506,6 +509,7 @@ fn test_tuple_type() {
                 module: "gleam".into(),
                 publicity: crate::ast::Publicity::Public,
                 package: "".into(),
+                narrowed_variant: None,
             }),
             Arc::new(Type::Named {
                 name: "String".into(),
@@ -513,6 +517,7 @@ fn test_tuple_type() {
                 module: "gleam".into(),
                 publicity: crate::ast::Publicity::Public,
                 package: "".into(),
+                narrowed_variant: None,
             }),
         ],
     };
@@ -535,6 +540,7 @@ fn test_fn_type() {
                 module: "gleam".into(),
                 publicity: crate::ast::Publicity::Public,
                 package: "".into(),
+                narrowed_variant: None,
             }),
             Arc::new(Type::Named {
                 name: "String".into(),
@@ -542,6 +548,7 @@ fn test_fn_type() {
                 module: "gleam".into(),
                 publicity: crate::ast::Publicity::Public,
                 package: "".into(),
+                narrowed_variant: None,
             }),
         ],
         retrn: Arc::new(Type::Named {
@@ -550,6 +557,7 @@ fn test_fn_type() {
             module: "gleam".into(),
             publicity: crate::ast::Publicity::Public,
             package: "".into(),
+            narrowed_variant: None,
         }),
     };
 
@@ -568,6 +576,7 @@ fn test_module_alias() {
         module: "mod1".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "animals.Cat");
@@ -591,6 +600,7 @@ fn test_type_alias_and_generics() {
         module: "mod".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "Cat(one)");
@@ -614,6 +624,7 @@ fn test_unqualified_import_and_generic() {
         module: "mod".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "C(one)");
@@ -629,6 +640,7 @@ fn nested_module() {
         module: "one/two/three".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "three.Cat");
@@ -652,6 +664,7 @@ fn test_unqualified_import_and_module_alias() {
         module: "mod1".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "C");
@@ -673,6 +686,7 @@ fn test_module_imports() {
         module: "mod".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     let typ1 = Type::Named {
@@ -681,6 +695,7 @@ fn test_module_imports() {
         module: "mod2".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     assert_eq!(printer.print_type(&type_), "animals.Cat");
@@ -709,6 +724,7 @@ fn test_multiple_generic_annotations() {
         module: "tigermodule".into(),
         publicity: crate::ast::Publicity::Public,
         package: "".into(),
+        narrowed_variant: None,
     };
 
     let typ1 = Type::Var {